int Rectangle::length => float Rectangle::length

+++ src/Rectangle.h
@@ -36,6 +36,6 @@ private:
 	
 	/* Le nombre d'arêtes. 0 pour un cercle normale (3 triangle ...)
 	*/
-	int length;
+	float length;
 
 };
